added home route

+++ app/Http/Controllers/Topic/CommentController.php
@@ -12,9 +12,18 @@ use App\Models\TopicCommentFile;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use App\Services\TopicCommentService;
 
 class CommentController extends Controller
 {
+    private TopicCommentService $topicCommentService;
+    private const TOPIC_COMMENT_FILES_PATH = 'topic';
+
+    public function __construct(TopicCommentService $topicCommentService)
+    {
+        $this->topicCommentService = $topicCommentService;
+    }
+
     public function store(StoreTopicComment $request)
     {
         $topicCommentId = TopicComment::insertGetId([
@@ -24,8 +33,9 @@ class CommentController extends Controller
             'created_at' => Carbon::now(),
         ]);
 
-        foreach($request->file('files') ?? [] as $file) {
-            $path = $file->store('topic-comment');
+        foreach($request->file('files') ?? [] as $file)
+        {
+            $path = $file->store(self::TOPIC_COMMENT_FILES_PATH);
 
             TopicCommentFile::create([
                 'topic_comment_id' => $topicCommentId,
@@ -42,7 +52,11 @@ class CommentController extends Controller
     {
         $topicComment = TopicComment::find($id);
 
-        $topicComment->delete();
+        if($topicComment)
+        {
+            $this->topicCommentService->destroyFiles($topicComment);
+            $topicComment->delete();
+        }
 
         return redirect()
             ->route('topic.get', ['id' => $topicComment->topic->id])
